move mediaplayer back into core package
