+++ app/src/main/java/com/guardian/podxdemo/dagger/AppComponent.kt
@@ -1,22 +1,22 @@
 package com.guardian.podxdemo.dagger
 
 import com.guardian.core.dagger.CoreComponent
-import com.guardian.mediaplayer.media.dagger.MediaPlayerComponent
 import com.guardian.podxdemo.PodXDemoApplication
 import dagger.Component
 import dagger.android.AndroidInjectionModule
+import javax.inject.Singleton
 
+@Singleton
 @Component(modules = [AndroidInjectionModule::class,
     MainActivityModule::class,
     PodcastFragmentsModule::class,
     PodcastViewModelsModule::class,
     AppExecutorsModule::class],
-    dependencies = [CoreComponent::class, MediaPlayerComponent::class])
+    dependencies = [CoreComponent::class])
 interface AppComponent {
     @Component.Builder
     interface Builder {
         fun coreComponent(component: CoreComponent): Builder
-        fun mediaPlayerComponent(component: MediaPlayerComponent): Builder
         fun build(): AppComponent
     }
