+++ app/src/main/java/com/guardian/podxdemo/presentation/feed/FeedViewModel.kt
@@ -1,16 +1,22 @@
 package com.guardian.podxdemo.presentation.feed
 
+import android.support.v4.media.MediaBrowserCompat
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.guardian.core.feed.Feed
 import com.guardian.core.feed.FeedRepository
 import com.guardian.core.search.SearchResult
+import com.guardian.core.mediaplayer.common.MediaSessionConnection
 import kotlinx.coroutines.launch
+import timber.log.Timber
 import javax.inject.Inject
 
 class FeedViewModel
-@Inject constructor(val feedRepository: FeedRepository) :
+@Inject constructor(
+    val feedRepository: FeedRepository,
+    val mediaSessionConnection: MediaSessionConnection
+) :
     ViewModel() {
     val feedData: MutableLiveData<Feed> = MutableLiveData(
         Feed("", "", "", "", listOf())
@@ -28,7 +34,21 @@ class FeedViewModel
         )
     }
 
-    fun getFeed(feedUrl: String) = viewModelScope.launch {
-        feedData.postValue(feedRepository.getFeed(feedUrl))
+    fun getFeed(feedUrl: String) {
+        mediaSessionConnection.subscribe(feedUrl, subscriptionCallback)
+
+        viewModelScope.launch {
+            feedData.postValue(feedRepository.getFeed(feedUrl))
+        }
+    }
+
+    object subscriptionCallback : MediaBrowserCompat.SubscriptionCallback() {
+        override fun onChildrenLoaded(
+            parentId: String,
+            children: MutableList<MediaBrowserCompat.MediaItem>
+        ) {
+            children.forEach { Timber.i(it.mediaId) }
+            super.onChildrenLoaded(parentId, children)
+        }
     }
 }
