+++ core/src/main/java/com/guardian/core/audio/AudioRepository.kt
@@ -1,0 +1,3 @@
+package com.guardian.core.audio
+
+interface AudioRepository
