+++ core/src/main/java/com/guardian/core/feed/api/FeedItemXmlDataObject.kt
@@ -9,7 +9,8 @@ data class FeedItemXmlDataObject(
     val description: String = "",
     val itunesImage: FeedItunesImageXmlDataObject = FeedItunesImageXmlDataObject(),
     val image: FeedImageXmlDataObject = FeedImageXmlDataObject(),
-    val pubDate: String = ""
+    val pubDate: String = "",
+    val enclosureXmlDataObject: FeedItemEnclosureXmlDataObject = FeedItemEnclosureXmlDataObject()
 ) : XmlDataObject {
     override fun isEmpty(): Boolean = title.isEmpty()
 
@@ -20,21 +21,23 @@ data class FeedItemXmlDataObject(
             return mutableMapOf(
                 "title" to ValueContainer(""),
                 "description" to ValueContainer(""),
-                "itunes:image" to ValueContainer(
-                    FeedItunesImageXmlDataObject()
-                ),
+                "itunes:image" to ValueContainer(FeedItunesImageXmlDataObject()),
                 "image" to ValueContainer(FeedImageXmlDataObject()),
-                "pubDate" to ValueContainer("")
+                "pubDate" to ValueContainer(""),
+                "enclosure" to ValueContainer(FeedItemEnclosureXmlDataObject())
             )
         }
 
         override fun instantiateFromXmlParserElementMap(xmlParserElementMap: Map<String, ValueContainer<*>>): XmlDataObject {
             return FeedItemXmlDataObject(
-                xmlParserElementMap["title"]?.value as String,
-                xmlParserElementMap["description"]?.value as String,
-                xmlParserElementMap["itunes:image"]?.value as FeedItunesImageXmlDataObject,
-                xmlParserElementMap["image"]?.value as FeedImageXmlDataObject,
-                xmlParserElementMap["pubDate"]?.value as String
+                title = xmlParserElementMap["title"]?.value as String,
+                description = xmlParserElementMap["description"]?.value as String,
+                itunesImage = xmlParserElementMap["itunes:image"]?.value
+                    as FeedItunesImageXmlDataObject,
+                image = xmlParserElementMap["image"]?.value as FeedImageXmlDataObject,
+                pubDate = xmlParserElementMap["pubDate"]?.value as String,
+                enclosureXmlDataObject = xmlParserElementMap["enclosure"]?.value
+                    as FeedItemEnclosureXmlDataObject
             )
         }
     }
