+++ core/src/main/java/com/guardian/core/mediaplayer/MediaService.kt
@@ -1,0 +1,482 @@
+package com.guardian.core.mediaplayer
+
+import android.app.PendingIntent
+import android.app.Service
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+import android.content.IntentFilter
+import android.media.AudioManager
+import android.net.Uri
+import android.os.Build
+import android.os.Bundle
+import android.support.v4.media.MediaBrowserCompat
+import android.support.v4.media.MediaBrowserCompat.MediaItem
+import android.support.v4.media.MediaDescriptionCompat
+import android.support.v4.media.MediaMetadataCompat
+import android.support.v4.media.session.MediaControllerCompat
+import android.support.v4.media.session.MediaSessionCompat
+import android.support.v4.media.session.PlaybackStateCompat
+import androidx.core.app.NotificationManagerCompat
+import androidx.core.content.ContextCompat
+import androidx.media.MediaBrowserServiceCompat
+import com.google.android.exoplayer2.C
+import com.google.android.exoplayer2.ExoPlayer
+import com.google.android.exoplayer2.ExoPlayerFactory
+import com.google.android.exoplayer2.Player
+import com.google.android.exoplayer2.Timeline
+import com.google.android.exoplayer2.audio.AudioAttributes
+import com.google.android.exoplayer2.ext.mediasession.MediaSessionConnector
+import com.google.android.exoplayer2.ext.mediasession.TimelineQueueNavigator
+import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory
+import com.google.android.exoplayer2.util.Util
+import com.guardian.core.R
+import com.guardian.core.mediaplayer.extensions.flag
+import com.guardian.core.mediaplayer.library.BrowseTree
+import com.guardian.core.mediaplayer.library.JsonSource
+import com.guardian.core.mediaplayer.library.MEDIA_SEARCH_SUPPORTED
+import com.guardian.core.mediaplayer.library.MusicSource
+import com.guardian.core.mediaplayer.library.UAMP_BROWSABLE_ROOT
+import com.guardian.core.mediaplayer.library.UAMP_EMPTY_ROOT
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.launch
+
+/**
+ * Google's UAMP media service to be repurposed for Podex
+ *
+ *
+ * This class is the entry point for browsing and playback commands from the APP's UI
+ * and other apps that wish to play music via UAMP (for example, Android Auto or
+ * the Google Assistant).
+ *
+ * Browsing begins with the method [MediaService.onGetRoot], and continues in
+ * the callback [MediaService.onLoadChildren].
+ *
+ * For more information on implementing a MediaBrowserService,
+ * visit [https://developer.android.com/guide/topics/media-apps/audio-app/building-a-mediabrowserservice.html](https://developer.android.com/guide/topics/media-apps/audio-app/building-a-mediabrowserservice.html).
+ */
+open class MediaService : MediaBrowserServiceCompat() {
+    /*
+    * todo
+    *  val browseTree mutable with lists of search results
+    *  fun add browsable
+    *  callback metadata for id
+    */
+
+    private lateinit var becomingNoisyReceiver: BecomingNoisyReceiver
+    private lateinit var notificationManager: NotificationManagerCompat
+    private lateinit var notificationBuilder: NotificationBuilder
+    private lateinit var mediaSource: MusicSource
+    private lateinit var packageValidator: PackageValidator
+
+    private val serviceJob = SupervisorJob()
+    private val serviceScope = CoroutineScope(Dispatchers.Main + serviceJob)
+
+    protected lateinit var mediaSession: MediaSessionCompat
+    protected lateinit var mediaController: MediaControllerCompat
+    protected lateinit var mediaSessionConnector: MediaSessionConnector
+
+    /**
+     * This must be `by lazy` because the source won't initially be ready.
+     * See [MediaService.onLoadChildren] to see where it's accessed (and first
+     * constructed).
+     */
+    private val browseTree: BrowseTree by lazy {
+        BrowseTree(mediaSource)
+    }
+
+    private var isForegroundService = false
+
+    private val remoteJsonSource: Uri =
+        Uri.parse("https://storage.googleapis.com/uamp/catalog.json")
+
+    private val uAmpAudioAttributes = AudioAttributes.Builder()
+        .setContentType(C.CONTENT_TYPE_MUSIC)
+        .setUsage(C.USAGE_MEDIA)
+        .build()
+
+    /**
+     * Configure ExoPlayer to handle audio focus for us.
+     * See [Player.AudioComponent.setAudioAttributes] for details.
+     */
+    private val exoPlayer: ExoPlayer by lazy {
+        ExoPlayerFactory.newSimpleInstance(this).apply {
+            setAudioAttributes(uAmpAudioAttributes, true)
+        }
+    }
+
+    @ExperimentalCoroutinesApi
+    override fun onCreate() {
+        super.onCreate()
+
+        // Build a PendingIntent that can be used to launch the UI.
+        val sessionActivityPendingIntent =
+            packageManager?.getLaunchIntentForPackage(packageName)?.let { sessionIntent ->
+                PendingIntent.getActivity(this, 0, sessionIntent, 0)
+            }
+
+        // Create a new MediaSession.
+        mediaSession = MediaSessionCompat(this, "MediaService")
+            .apply {
+                setSessionActivity(sessionActivityPendingIntent)
+                isActive = true
+            }
+
+        /**
+         * In order for [MediaBrowserCompat.ConnectionCallback.onConnected] to be called,
+         * a [MediaSessionCompat.Token] needs to be set on the [MediaBrowserServiceCompat].
+         *
+         * It is possible to wait to set the session token, if required for a specific use-case.
+         * However, the token *must* be set by the time [MediaBrowserServiceCompat.onGetRoot]
+         * returns, or the connection will fail silently. (The system will not even call
+         * [MediaBrowserCompat.ConnectionCallback.onConnectionFailed].)
+         */
+        sessionToken = mediaSession.sessionToken
+
+        // Because ExoPlayer will manage the MediaSession, add the service as a callback for
+        // state changes.
+        mediaController = MediaControllerCompat(this, mediaSession).also {
+            it.registerCallback(MediaControllerCallback())
+        }
+
+        notificationBuilder = NotificationBuilder(this)
+        notificationManager = NotificationManagerCompat.from(this)
+
+        becomingNoisyReceiver =
+            BecomingNoisyReceiver(
+                context = this,
+                sessionToken = mediaSession.sessionToken
+            )
+
+        // The media library is built from a remote JSON file. We'll create the source here,
+        // and then use a suspend function to perform the download off the main thread.
+        mediaSource = JsonSource(
+            context = this,
+            source = remoteJsonSource
+        )
+        serviceScope.launch {
+            mediaSource.load()
+        }
+
+        // ExoPlayer will manage the MediaSession for us.
+        mediaSessionConnector = MediaSessionConnector(mediaSession).also { connector ->
+            // Produces DataSource instances through which media data is loaded.
+            val dataSourceFactory = DefaultDataSourceFactory(
+                this, Util.getUserAgent(this, UAMP_USER_AGENT), null
+            )
+
+            // Create the PlaybackPreparer of the media session connector.
+            val playbackPreparer = UampPlaybackPreparer(
+                mediaSource,
+                exoPlayer,
+                dataSourceFactory
+            )
+
+            connector.setPlayer(exoPlayer)
+            connector.setPlaybackPreparer(playbackPreparer)
+            connector.setQueueNavigator(
+                UampQueueNavigator(
+                    mediaSession
+                )
+            )
+        }
+
+        packageValidator = PackageValidator(
+            this,
+            R.xml.allowed_media_browser_callers
+        )
+    }
+
+    /**
+     * This is the code that causes UAMP to stop playing when swiping it away from recents.
+     * The choice to do this is app specific. Some apps stop playback, while others allow playback
+     * to continue and allow uses to stop it with the notification.
+     */
+    override fun onTaskRemoved(rootIntent: Intent) {
+        super.onTaskRemoved(rootIntent)
+
+        /**
+         * By stopping playback, the player will transition to [Player.STATE_IDLE]. This will
+         * cause a state change in the MediaSession, and (most importantly) call
+         * [MediaControllerCallback.onPlaybackStateChanged]. Because the playback state will
+         * be reported as [PlaybackStateCompat.STATE_NONE], the service will first remove
+         * itself as a foreground service, and will then call [stopSelf].
+         */
+        exoPlayer.stop(true)
+    }
+
+    override fun onDestroy() {
+        mediaSession.run {
+            isActive = false
+            release()
+        }
+
+        // Cancel coroutines when the service is going away.
+        serviceJob.cancel()
+    }
+
+    /**
+     * Returns the "root" media ID that the client should request to get the list of
+     * [MediaItem]s to browse/play.
+     */
+    override fun onGetRoot(
+        clientPackageName: String,
+        clientUid: Int,
+        rootHints: Bundle?
+    ): BrowserRoot? {
+
+        /*
+         * By default, all known clients are permitted to search, but only tell unknown callers
+         * about search if permitted by the [BrowseTree].
+         */
+        val isKnownCaller = packageValidator.isKnownCaller(clientPackageName, clientUid)
+        val rootExtras = Bundle().apply {
+            putBoolean(
+                MEDIA_SEARCH_SUPPORTED,
+                isKnownCaller || browseTree.searchableByUnknownCaller
+            )
+            putBoolean(CONTENT_STYLE_SUPPORTED, true)
+            putInt(
+                CONTENT_STYLE_BROWSABLE_HINT,
+                CONTENT_STYLE_GRID
+            )
+            putInt(
+                CONTENT_STYLE_PLAYABLE_HINT,
+                CONTENT_STYLE_LIST
+            )
+        }
+
+        return if (true) { // isKnownCaller) {
+            // The caller is allowed to browse, so return the root.
+            BrowserRoot(UAMP_BROWSABLE_ROOT, rootExtras)
+        } else {
+            /**
+             * Unknown caller. There are two main ways to handle this:
+             * 1) Return a root without any content, which still allows the connecting client
+             * to issue commands.
+             * 2) Return `null`, which will cause the system to disconnect the app.
+             *
+             * UAMP takes the first approach for a variety of reasons, but both are valid
+             * options.
+             */
+            BrowserRoot(UAMP_EMPTY_ROOT, rootExtras)
+        }
+    }
+
+    /**
+     * Returns (via the [result] parameter) a list of [MediaItem]s that are child
+     * items of the provided [parentMediaId]. See [BrowseTree] for more details on
+     * how this is build/more details about the relationships.
+     */
+    override fun onLoadChildren(
+        parentMediaId: String,
+        result: Result<List<MediaItem>>
+    ) {
+
+        // If the media source is ready, the results will be set synchronously here.
+        val resultsSent = mediaSource.whenReady { successfullyInitialized ->
+            if (successfullyInitialized) {
+                val children = browseTree[parentMediaId]?.map { item ->
+                    MediaItem(item.description, item.flag)
+                }
+                result.sendResult(children)
+            } else {
+                mediaSession.sendSessionEvent(NETWORK_FAILURE, null)
+                result.sendResult(null)
+            }
+        }
+
+        // If the results are not ready, the service must "detach" the results before
+        // the method returns. After the source is ready, the lambda above will run,
+        // and the caller will be notified that the results are ready.
+        //
+        // See [MediaItemFragmentViewModel.subscriptionCallback] for how this is passed to the
+        // UI/displayed in the [RecyclerView].
+        if (!resultsSent) {
+            result.detach()
+        }
+    }
+
+    /**
+     * Returns a list of [MediaItem]s that match the given search query
+     */
+    override fun onSearch(
+        query: String,
+        extras: Bundle?,
+        result: Result<List<MediaItem>>
+    ) {
+
+        val resultsSent = mediaSource.whenReady { successfullyInitialized ->
+            if (successfullyInitialized) {
+                val resultsList = mediaSource.search(query, extras ?: Bundle.EMPTY)
+                    .map { mediaMetadata ->
+                        MediaItem(mediaMetadata.description, mediaMetadata.flag)
+                    }
+                result.sendResult(resultsList)
+            }
+        }
+
+        if (!resultsSent) {
+            result.detach()
+        }
+    }
+
+    /**
+     * Removes the [NOW_PLAYING_NOTIFICATION] notification.
+     *
+     * Since `stopForeground(false)` was already called (see
+     * [MediaControllerCallback.onPlaybackStateChanged], it's possible to cancel the notification
+     * with `notificationManager.cancel(NOW_PLAYING_NOTIFICATION)` if minSdkVersion is >=
+     * [Build.VERSION_CODES.LOLLIPOP].
+     *
+     * Prior to [Build.VERSION_CODES.LOLLIPOP], notifications associated with a foreground
+     * service remained marked as "ongoing" even after calling [Service.stopForeground],
+     * and cannot be cancelled normally.
+     *
+     * Fortunately, it's possible to simply call [Service.stopForeground] a second time, this
+     * time with `true`. This won't change anything about the service's state, but will simply
+     * remove the notification.
+     */
+    private fun removeNowPlayingNotification() {
+        stopForeground(true)
+    }
+
+    /**
+     * Class to receive callbacks about state changes to the [MediaSessionCompat]. In response
+     * to those callbacks, this class:
+     *
+     * - Build/update the service's notification.
+     * - Register/unregister a broadcast receiver for [AudioManager.ACTION_AUDIO_BECOMING_NOISY].
+     * - Calls [Service.startForeground] and [Service.stopForeground].
+     */
+    private inner class MediaControllerCallback : MediaControllerCompat.Callback() {
+        override fun onMetadataChanged(metadata: MediaMetadataCompat?) {
+            mediaController.playbackState?.let { updateNotification(it) }
+        }
+
+        override fun onPlaybackStateChanged(state: PlaybackStateCompat?) {
+            state?.let { updateNotification(it) }
+        }
+
+        private fun updateNotification(state: PlaybackStateCompat) {
+            val updatedState = state.state
+
+            // Skip building a notification when state is "none" and metadata is null.
+            val notification = if (mediaController.metadata != null &&
+                updatedState != PlaybackStateCompat.STATE_NONE) {
+                notificationBuilder.buildNotification(mediaSession.sessionToken)
+            } else {
+                null
+            }
+
+            when (updatedState) {
+                PlaybackStateCompat.STATE_BUFFERING,
+                PlaybackStateCompat.STATE_PLAYING -> {
+                    becomingNoisyReceiver.register()
+
+                    /**
+                     * This may look strange, but the documentation for [Service.startForeground]
+                     * notes that "calling this method does *not* put the service in the started
+                     * state itself, even though the name sounds like it."
+                     */
+                    if (notification != null) {
+                        notificationManager.notify(NOW_PLAYING_NOTIFICATION, notification)
+
+                        if (!isForegroundService) {
+                            ContextCompat.startForegroundService(
+                                applicationContext,
+                                Intent(applicationContext, this@MediaService.javaClass)
+                            )
+                            startForeground(NOW_PLAYING_NOTIFICATION, notification)
+                            isForegroundService = true
+                        }
+                    }
+                }
+                else -> {
+                    becomingNoisyReceiver.unregister()
+
+                    if (isForegroundService) {
+                        stopForeground(false)
+                        isForegroundService = false
+
+                        // If playback has ended, also stop the service.
+                        if (updatedState == PlaybackStateCompat.STATE_NONE) {
+                            stopSelf()
+                        }
+
+                        if (notification != null) {
+                            notificationManager.notify(NOW_PLAYING_NOTIFICATION, notification)
+                        } else {
+                            removeNowPlayingNotification()
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+/**
+ * Helper class to retrieve the the Metadata necessary for the ExoPlayer MediaSession connection
+ * extension to call [MediaSessionCompat.setMetadata].
+ */
+private class UampQueueNavigator(
+    mediaSession: MediaSessionCompat
+) : TimelineQueueNavigator(mediaSession) {
+    private val window = Timeline.Window()
+    override fun getMediaDescription(player: Player, windowIndex: Int): MediaDescriptionCompat =
+        player.currentTimeline
+            .getWindow(windowIndex, window, true).tag as MediaDescriptionCompat
+}
+
+/**
+ * Helper class for listening for when headphones are unplugged (or the audio
+ * will otherwise cause playback to become "noisy").
+ */
+private class BecomingNoisyReceiver(
+    private val context: Context,
+    sessionToken: MediaSessionCompat.Token
+) : BroadcastReceiver() {
+
+    private val noisyIntentFilter = IntentFilter(AudioManager.ACTION_AUDIO_BECOMING_NOISY)
+    private val controller = MediaControllerCompat(context, sessionToken)
+
+    private var registered = false
+
+    fun register() {
+        if (!registered) {
+            context.registerReceiver(this, noisyIntentFilter)
+            registered = true
+        }
+    }
+
+    fun unregister() {
+        if (registered) {
+            context.unregisterReceiver(this)
+            registered = false
+        }
+    }
+
+    override fun onReceive(context: Context, intent: Intent) {
+        if (intent.action == AudioManager.ACTION_AUDIO_BECOMING_NOISY) {
+            controller.transportControls.pause()
+        }
+    }
+}
+
+/*
+ * (Media) Session events
+ */
+const val NETWORK_FAILURE = "com.guardian.mediaplayer.media.session.NETWORK_FAILURE"
+
+/** Content styling constants */
+private const val CONTENT_STYLE_BROWSABLE_HINT = "android.media.browse.CONTENT_STYLE_BROWSABLE_HINT"
+private const val CONTENT_STYLE_PLAYABLE_HINT = "android.media.browse.CONTENT_STYLE_PLAYABLE_HINT"
+private const val CONTENT_STYLE_SUPPORTED = "android.media.browse.CONTENT_STYLE_SUPPORTED"
+private const val CONTENT_STYLE_LIST = 1
+private const val CONTENT_STYLE_GRID = 2
+
+private const val UAMP_USER_AGENT = "uamp.next"
