+++ core/src/main/java/com/guardian/core/mediaplayer/common/MediaSessionConnection.kt
@@ -1,0 +1,189 @@
+/*
+ * Copyright 2018 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.guardian.core.mediaplayer.common
+
+import android.content.ComponentName
+import android.content.Context
+import android.os.Bundle
+import android.os.Handler
+import android.os.ResultReceiver
+import android.support.v4.media.MediaBrowserCompat
+import android.support.v4.media.MediaMetadataCompat
+import android.support.v4.media.session.MediaControllerCompat
+import android.support.v4.media.session.MediaSessionCompat
+import android.support.v4.media.session.PlaybackStateCompat
+import androidx.lifecycle.MutableLiveData
+import androidx.media.MediaBrowserServiceCompat
+import com.guardian.core.mediaplayer.NETWORK_FAILURE
+import com.guardian.core.mediaplayer.common.MediaSessionConnection.MediaBrowserConnectionCallback
+
+/**
+ * Class that manages a connection to a [MediaBrowserServiceCompat] instance.
+ *
+ * Typically it's best to construct/inject dependencies either using DI or, as UAMP does,
+ * using [InjectorUtils] in the app module. There are a few difficulties for that here:
+ * - [MediaBrowserCompat] is a final class, so mocking it directly is difficult.
+ * - A [MediaBrowserConnectionCallback] is a parameter into the construction of
+ *   a [MediaBrowserCompat], and provides callbacks to this class.
+ * - [MediaBrowserCompat.ConnectionCallback.onConnected] is the best place to construct
+ *   a [MediaControllerCompat] that will be used to control the [MediaSessionCompat].
+ *
+ *  Because of these reasons, rather than constructing additional classes, this is treated as
+ *  a black box (which is why there's very little logic here).
+ *
+ *  This is also why the parameters to construct a [MediaSessionConnection] are simple
+ *  parameters, rather than private properties. They're only required to build the
+ *  [MediaBrowserConnectionCallback] and [MediaBrowserCompat] objects.
+ */
+class MediaSessionConnection(context: Context, serviceComponent: ComponentName) {
+    val isConnected = MutableLiveData<Boolean>()
+        .apply { postValue(false) }
+    val networkFailure = MutableLiveData<Boolean>()
+        .apply { postValue(false) }
+
+    val rootMediaId: String get() = mediaBrowser.root
+
+    val playbackState = MutableLiveData<PlaybackStateCompat>()
+        .apply { postValue(EMPTY_PLAYBACK_STATE) }
+    val nowPlaying = MutableLiveData<MediaMetadataCompat>()
+        .apply { postValue(NOTHING_PLAYING) }
+
+    val transportControls: MediaControllerCompat.TransportControls
+        get() = mediaController.transportControls
+
+    private val mediaBrowserConnectionCallback = MediaBrowserConnectionCallback(context)
+    private val mediaBrowser = MediaBrowserCompat(
+        context,
+        serviceComponent,
+        mediaBrowserConnectionCallback, null
+    ).apply { connect() }
+    private lateinit var mediaController: MediaControllerCompat
+
+    fun subscribe(parentId: String, callback: MediaBrowserCompat.SubscriptionCallback) {
+        mediaBrowser.subscribe(parentId, callback)
+    }
+
+    fun unsubscribe(parentId: String, callback: MediaBrowserCompat.SubscriptionCallback) {
+        mediaBrowser.unsubscribe(parentId, callback)
+    }
+
+    fun sendCommand(command: String, parameters: Bundle?) =
+            sendCommand(command, parameters) { _, _ -> }
+
+    fun sendCommand(
+        command: String,
+        parameters: Bundle?,
+        resultCallback: ((Int, Bundle?) -> Unit)
+    ) = if (mediaBrowser.isConnected) {
+        mediaController.sendCommand(command, parameters, object : ResultReceiver(Handler()) {
+            override fun onReceiveResult(resultCode: Int, resultData: Bundle?) {
+                resultCallback(resultCode, resultData)
+            }
+        })
+        true
+    } else {
+        false
+    }
+
+    private inner class MediaBrowserConnectionCallback(private val context: Context) :
+        MediaBrowserCompat.ConnectionCallback() {
+        /**
+         * Invoked after [MediaBrowserCompat.connect] when the request has successfully
+         * completed.
+         */
+        override fun onConnected() {
+            // Get a MediaController for the MediaSession.
+            mediaController = MediaControllerCompat(context, mediaBrowser.sessionToken).apply {
+                registerCallback(MediaControllerCallback())
+            }
+
+            isConnected.postValue(true)
+        }
+
+        /**
+         * Invoked when the client is disconnected from the media browser.
+         */
+        override fun onConnectionSuspended() {
+            isConnected.postValue(false)
+        }
+
+        /**
+         * Invoked when the connection to the media browser failed.
+         */
+        override fun onConnectionFailed() {
+            isConnected.postValue(false)
+        }
+    }
+
+    private inner class MediaControllerCallback : MediaControllerCompat.Callback() {
+
+        override fun onPlaybackStateChanged(state: PlaybackStateCompat?) {
+            playbackState.postValue(state ?: EMPTY_PLAYBACK_STATE)
+        }
+
+        override fun onMetadataChanged(metadata: MediaMetadataCompat?) {
+            nowPlaying.postValue(metadata ?: NOTHING_PLAYING)
+        }
+
+        override fun onQueueChanged(queue: MutableList<MediaSessionCompat.QueueItem>?) {
+        }
+
+        override fun onSessionEvent(event: String?, extras: Bundle?) {
+            super.onSessionEvent(event, extras)
+            when (event) {
+                NETWORK_FAILURE -> networkFailure.postValue(true)
+            }
+        }
+
+        /**
+         * Normally if a [MediaBrowserServiceCompat] drops its connection the callback comes via
+         * [MediaControllerCompat.Callback] (here). But since other connection status events
+         * are sent to [MediaBrowserCompat.ConnectionCallback], we catch the disconnect here and
+         * send it on to the other callback.
+         */
+        override fun onSessionDestroyed() {
+            mediaBrowserConnectionCallback.onConnectionSuspended()
+        }
+    }
+
+    companion object {
+        // For Singleton instantiation.
+        @Volatile
+        private var instance: MediaSessionConnection? = null
+
+        fun getInstance(context: Context, serviceComponent: ComponentName) =
+            instance ?: synchronized(this) {
+                instance
+                    ?: MediaSessionConnection(
+                        context,
+                        serviceComponent
+                    )
+                    .also { instance = it }
+            }
+    }
+}
+
+@Suppress("PropertyName")
+val EMPTY_PLAYBACK_STATE: PlaybackStateCompat = PlaybackStateCompat.Builder()
+    .setState(PlaybackStateCompat.STATE_NONE, 0, 0f)
+    .build()
+
+@Suppress("PropertyName")
+val NOTHING_PLAYING: MediaMetadataCompat = MediaMetadataCompat.Builder()
+    .putString(MediaMetadataCompat.METADATA_KEY_MEDIA_ID, "")
+    .putLong(MediaMetadataCompat.METADATA_KEY_DURATION, 0)
+    .build()
