+++ core/src/main/java/com/guardian/core/mediaplayer/library/JsonSource.kt
@@ -1,0 +1,235 @@
+/*
+ * Copyright 2017 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.guardian.core.mediaplayer.library
+
+import android.content.Context
+import android.net.Uri
+import android.support.v4.media.MediaBrowserCompat.MediaItem
+import android.support.v4.media.MediaDescriptionCompat.STATUS_NOT_DOWNLOADED
+import android.support.v4.media.MediaMetadataCompat
+import com.bumptech.glide.Glide
+import com.bumptech.glide.RequestManager
+import com.bumptech.glide.load.engine.DiskCacheStrategy
+import com.bumptech.glide.request.RequestOptions
+import com.google.gson.Gson
+import com.guardian.core.mediaplayer.extensions.album
+import com.guardian.core.mediaplayer.extensions.albumArt
+import com.guardian.core.mediaplayer.extensions.albumArtUri
+import com.guardian.core.mediaplayer.extensions.artist
+import com.guardian.core.mediaplayer.extensions.displayDescription
+import com.guardian.core.mediaplayer.extensions.displayIconUri
+import com.guardian.core.mediaplayer.extensions.displaySubtitle
+import com.guardian.core.mediaplayer.extensions.displayTitle
+import com.guardian.core.mediaplayer.extensions.downloadStatus
+import com.guardian.core.mediaplayer.extensions.duration
+import com.guardian.core.mediaplayer.extensions.flag
+import com.guardian.core.mediaplayer.extensions.genre
+import com.guardian.core.mediaplayer.extensions.id
+import com.guardian.core.mediaplayer.extensions.mediaUri
+import com.guardian.core.mediaplayer.extensions.title
+import com.guardian.core.mediaplayer.extensions.trackCount
+import com.guardian.core.mediaplayer.extensions.trackNumber
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import java.io.BufferedReader
+import java.io.IOException
+import java.io.InputStreamReader
+import java.net.URL
+import java.util.concurrent.TimeUnit
+
+/**
+ * Source of [MediaMetadataCompat] objects created from a basic JSON stream.
+ *
+ * The definition of the JSON is specified in the docs of [JsonMusic] in this file,
+ * which is the object representation of it.
+ */
+class JsonSource(context: Context, private val source: Uri) : AbstractMusicSource() {
+
+    private var catalog: List<MediaMetadataCompat> = emptyList()
+    private val glide: RequestManager
+
+    init {
+        state = STATE_INITIALIZING
+        glide = Glide.with(context)
+    }
+
+    override fun iterator(): Iterator<MediaMetadataCompat> = catalog.iterator()
+
+    override suspend fun load() {
+        updateCatalog(source)?.let { updatedCatalog ->
+            catalog = updatedCatalog
+            state = STATE_INITIALIZED
+        } ?: run {
+            catalog = emptyList()
+            state = STATE_ERROR
+        }
+    }
+
+    /**
+     * Function to connect to a remote URI and download/process the JSON file that corresponds to
+     * [MediaMetadataCompat] objects.
+     */
+    private suspend fun updateCatalog(catalogUri: Uri): List<MediaMetadataCompat>? {
+        return withContext(Dispatchers.IO) {
+            val musicCat = try {
+                downloadJson(catalogUri)
+            } catch (ioException: IOException) {
+                return@withContext null
+            }
+
+            // Get the base URI to fix up relative references later.
+            val baseUri = catalogUri.toString().removeSuffix(catalogUri.lastPathSegment ?: "")
+
+            musicCat.music.map { song ->
+                // The JSON may have paths that are relative to the source of the JSON
+                // itself. We need to fix them up here to turn them into absolute paths.
+                catalogUri.scheme?.let { scheme ->
+                    if (!song.source.startsWith(scheme)) {
+                        song.source = baseUri + song.source
+                    }
+                    if (!song.image.startsWith(scheme)) {
+                        song.image = baseUri + song.image
+                    }
+                }
+
+                // Block on downloading artwork.
+                val art = glide.applyDefaultRequestOptions(glideOptions)
+                    .asBitmap()
+                    .load(song.image)
+                    .submit(
+                        NOTIFICATION_LARGE_ICON_SIZE,
+                        NOTIFICATION_LARGE_ICON_SIZE
+                    )
+                    .get()
+
+                MediaMetadataCompat.Builder()
+                    .from(song)
+                    .apply {
+                        albumArt = art
+                    }
+                    .build()
+            }.toList()
+        }
+    }
+
+    /**
+     * Attempts to download a catalog from a given Uri.
+     *
+     * @param catalogUri URI to attempt to download the catalog form.
+     * @return The catalog downloaded, or an empty catalog if an error occurred.
+     */
+    @Throws(IOException::class)
+    private fun downloadJson(catalogUri: Uri): JsonCatalog {
+        val catalogConn = URL(catalogUri.toString())
+        val reader = BufferedReader(InputStreamReader(catalogConn.openStream()))
+        return Gson().fromJson<JsonCatalog>(reader, JsonCatalog::class.java)
+    }
+}
+
+/**
+ * Extension method for [MediaMetadataCompat.Builder] to set the fields from
+ * our JSON constructed object (to make the code a bit easier to see).
+ */
+fun MediaMetadataCompat.Builder.from(jsonMusic: JsonMusic): MediaMetadataCompat.Builder {
+    // The duration from the JSON is given in seconds, but the rest of the code works in
+    // milliseconds. Here's where we convert to the proper units.
+    val durationMs = TimeUnit.SECONDS.toMillis(jsonMusic.duration)
+
+    id = jsonMusic.id
+    title = jsonMusic.title
+    artist = jsonMusic.artist
+    album = jsonMusic.album
+    duration = durationMs
+    genre = jsonMusic.genre
+    mediaUri = jsonMusic.source
+    albumArtUri = jsonMusic.image
+    trackNumber = jsonMusic.trackNumber
+    trackCount = jsonMusic.totalTrackCount
+    flag = MediaItem.FLAG_PLAYABLE
+
+    // To make things easier for *displaying* these, set the display properties as well.
+    displayTitle = jsonMusic.title
+    displaySubtitle = jsonMusic.artist
+    displayDescription = jsonMusic.album
+    displayIconUri = jsonMusic.image
+
+    // Add downloadStatus to force the creation of an "extras" bundle in the resulting
+    // MediaMetadataCompat object. This is needed to send accurate metadata to the
+    // media session during updates.
+    downloadStatus = STATUS_NOT_DOWNLOADED
+
+    // Allow it to be used in the typical builder style.
+    return this
+}
+
+/**
+ * Wrapper object for our JSON in order to be processed easily by GSON.
+ */
+class JsonCatalog {
+    var music: List<JsonMusic> = listOf()
+}
+
+/**
+ * An individual piece of music included in our JSON catalog.
+ * The format from the server is as specified:
+ * ```
+ *     { "music" : [
+ *     { "title" : // Title of the piece of music
+ *     "album" : // Album title of the piece of music
+ *     "artist" : // Artist of the piece of music
+ *     "genre" : // Primary genre of the music
+ *     "source" : // Path to the music, which may be relative
+ *     "image" : // Path to the art for the music, which may be relative
+ *     "trackNumber" : // Track number
+ *     "totalTrackCount" : // Track count
+ *     "duration" : // Duration of the music in seconds
+ *     "site" : // Source of the music, if applicable
+ *     }
+ *     ]}
+ * ```
+ *
+ * `source` and `image` can be provided in either relative or
+ * absolute paths. For example:
+ * ``
+ *     "source" : "https://www.example.com/music/ode_to_joy.mp3",
+ *     "image" : "ode_to_joy.jpg"
+ * ``
+ *
+ * The `source` specifies the full URI to download the piece of music from, but
+ * `image` will be fetched relative to the path of the JSON file itself. This means
+ * that if the JSON was at "https://www.example.com/json/music.json" then the image would be found
+ * at "https://www.example.com/json/ode_to_joy.jpg".
+ */
+@Suppress("unused")
+class JsonMusic {
+    var id: String = ""
+    var title: String = ""
+    var album: String = ""
+    var artist: String = ""
+    var genre: String = ""
+    var source: String = ""
+    var image: String = ""
+    var trackNumber: Long = 0
+    var totalTrackCount: Long = 0
+    var duration: Long = -1
+    var site: String = ""
+}
+
+private const val NOTIFICATION_LARGE_ICON_SIZE = 144 // px
+
+private val glideOptions = RequestOptions()
+    .diskCacheStrategy(DiskCacheStrategy.RESOURCE)
