+++ core/src/main/java/com/guardian/core/mediaplayer/library/RemoteSource.kt
@@ -1,0 +1,13 @@
+package com.guardian.core.mediaplayer.library
+
+import android.support.v4.media.MediaMetadataCompat
+
+class RemoteSource : AbstractMusicSource() {
+    override suspend fun load() {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun iterator(): Iterator<MediaMetadataCompat> {
+        TODO("not implemented") // To change body of created functions use File | Settings | File Templates.
+    }
+}
