--- mediaplayer/src/main/java/com/guardian/mediaplayer/media/extensions/JavaLangExt.kt
@@ -1,54 +1,0 @@
-/*
- * Copyright 2018 Google Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.guardian.mediaplayer.media.extensions
-
-import android.net.Uri
-import java.net.URLEncoder
-import java.nio.charset.Charset
-
-/**
- * This file contains extension methods for the java.lang package.
- */
-
-/**
- * Helper method to check if a [String] contains another in a case insensitive way.
- */
-fun String?.containsCaseInsensitive(other: String?) =
-        if (this == null && other == null) {
-            true
-        } else if (this != null && other != null) {
-            toLowerCase().contains(other.toLowerCase())
-        } else {
-            false
-        }
-
-/**
- * Helper extension to URL encode a [String]. Returns an empty string when called on null.
- */
-inline val String?.urlEncoded: String
-    get() = if (Charset.isSupported("UTF-8")) {
-        URLEncoder.encode(this ?: "", "UTF-8")
-    } else {
-        // If UTF-8 is not supported, use the default charset.
-        @Suppress("deprecation")
-        URLEncoder.encode(this ?: "")
-    }
-
-/**
- * Helper extension to convert a potentially null [String] to a [Uri] falling back to [Uri.EMPTY]
- */
-fun String?.toUri(): Uri = this?.let { Uri.parse(it) } ?: Uri.EMPTY
