--- mediaplayer/src/main/res/values/strings.xml
@@ -1,13 +1,0 @@
-<resources>
-    <string name="app_name">Podex:fundament</string>
-
-    <!-- Notificaiton Channel -->
-    <string name="notification_channel">Now Playing</string>
-    <string name="notification_channel_description">Shows what music is currently playing in UAMP</string>
-
-    <!-- Notification actions -->
-    <string name="notification_skip_to_previous">Skip to previous track</string>
-    <string name="notification_skip_to_next">Skip to next track</string>
-    <string name="notification_play">Play</string>
-    <string name="notification_pause">Pause</string>
-</resources>
